requirements.txt
numpy == 1.21.2
google-cloud-pubsub == 2.8.0